Add flows
